src/Python/pythonwebhost/requirements.txt
Django==2.2
aiohttp==3.5.4
alabaster==0.7.10
asn1crypto==0.24.0
astroid==1.5.3
astropy==2.0.2
async-timeout==3.0.1
attrs==19.1.0
autopep8==1.4.4
babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
bleach==2.0.0
bokeh==0.12.10
boto==2.48.0
Bottleneck==1.2.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.2
conda-verify==2.0.0
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
cycler==0.10.0
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.3
datashape==0.5.4
decorator==4.1.2
distlib==0.2.5
distributed==1.19.1
djangorestframework==3.9.3
django-rest-swagger==2.2.0
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.12
Flask==0.12.2
Flask-Cors==3.0.3
gevent==1.2.2
glob2==0.5
greenlet==0.4.12
h5py==2.7.0
heapdict==1.0.0
html5lib==0.999999999
idna==2.8
imageio==2.2.0
imagesize==0.7.1
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
itsdangerous==0.24
itypes==1.1.0
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lockfile==0.12.2
lxml==4.3.3
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
menuinst==1.4.10
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
multidict==4.5.2
multipledispatch==0.4.9
mypy==0.701
mypy-extensions==0.4.1
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nltk==3.2.4
nose==1.3.7
numba==0.35.0+10.g143f70e
numexpr==2.6.2
numpy==1.13.3
olefile==0.44
openapi-codec==1.3.2
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
partd==0.3.8
path.py==10.3.1
pathlib2==2.3.0
patsy==0.4.1
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.2.1
pip==19.0.3
pkginfo==1.4.1
ply==3.10
progress==1.3
prompt-toolkit==1.0.15
psutil==5.4.8
py==1.4.34
pycodestyle==2.5.0
pycosat==0.6.2
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.4
pyodbc==4.0.17
pyOpenSSL==19.0.0
pyparsing==2.2.0
PySocks==1.6.8
pytest==3.2.1
python-dateutil==2.6.1
pytz==2019.1
PyWavelets==0.5.2
pywin32==221
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
QtPy==1.3.1
requests==2.21.0
rope==0.10.5
ruamel-yaml==0.11.14
scikit-learn==0.19.1
scipy==0.19.1
seaborn==0.8
setuptools==41.0.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.1.13
sqlparse==0.3.0
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
testpath==0.3.1
toolz==0.8.2
tornado==4.5.2
traitlets==4.3.2
typed-ast==1.2.0
typing==3.6.4
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
wheel==0.33.1
wincertstore==0.2
win-inet-pton==1.1.0
win-unicode-console==0.5
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.4
xlwt==1.3.0
yarl==1.3.0
zict==0.1.3
